Changed License and minor fixes/improvements

+++ src/sampa/Core/MCache.php
@@ -4,11 +4,11 @@
 *	Memcache abstraction
 *
 *	@package sampa\Core\MCache
-*	@copyright 2013 appdeck
+*	@copyright 2016 appdeck
 *	@link http://github.com/appdeck/sampa
 *	@version 0.1
 *	@since 0.1
-*	@license http://www.gnu.org/licenses/gpl-3.0.html GNU General Public License, version 3
+*
 */
 
 namespace sampa\Core;
@@ -20,9 +20,11 @@ final class MCache {
 	private $memcache = null;
 
 	public function __construct($host, $port) {
-		$this->memcache = new \Memcache;
-		if (!$this->memcache->connect($host, $port))
-			throw new Exception\Cache("Can't connect to memcache server at {$host}:{$port}");
+		if (class_exists('\\Memcache')) {
+			$this->memcache = new \Memcache;
+			if (!$this->memcache->connect($host, $port))
+				throw new Exception\Cache("Can't connect to memcache server at {$host}:{$port}");
+		}
 	}
 
 	public static function singleton($host = 'localhost', $port = 11211) {
@@ -36,22 +38,32 @@ final class MCache {
 	}
 
 	public function flush() {
+		if (is_null($this->memcache))
+			return;
 		$this->memcache->flush();
 	}
 
 	public function extended_set($index, $value, $ttl) {
+		if (is_null($this->memcache))
+			return;
 		$this->memcache->set($index, $value, MEMCACHE_COMPRESSED, $ttl);
 	}
 
 	public function __set($index, $value) {
+		if (is_null($this->memcache))
+			return;
 		$this->memcache->set($index, $value, MEMCACHE_COMPRESSED, 0);
 	}
 
 	public function __get($index) {
+		if (is_null($this->memcache))
+			return null;
 		return $this->memcache->get($index);
 	}
 
 	public function __isset($index) {
+		if (is_null($this->memcache))
+			return false;
 		$flag = false;
 		if (($this->memcache->get($index, $flag) === false) && ($flag === false))
 			return false;
@@ -59,6 +71,8 @@ final class MCache {
 	}
 
 	public function __unset($index) {
+		if (is_null($this->memcache))
+			return;
 		$this->memcache->delete($index, 0);
 	}
 }
